+++ catalog/app/containers/Bucket/Search.js
@@ -6,10 +6,10 @@ import * as SearchResults from 'components/SearchResults'
 import * as AWS from 'utils/AWS'
 import * as BucketConfig from 'utils/BucketConfig'
 import * as Data from 'utils/Data'
+import MetaTitle from 'utils/MetaTitle'
 import * as NamedRoutes from 'utils/NamedRoutes'
 import parseSearch from 'utils/parseSearch'
 import search from 'utils/search'
-import mkStorage from 'utils/storage'
 import useEditableValue from 'utils/useEditableValue'
 
 function Browse({ bucket }) {
@@ -280,11 +280,6 @@ function Search({ bucket, query, page, mode, retry }) {
 
   const handleModeChange = React.useCallback(
     (newMode) => {
-      if (newMode === 'objects' || newMode === 'packages') {
-        storage.set('searchMode', newMode)
-      } else {
-        storage.remove('searchMode')
-      }
       history.push(urls.bucketSearch(bucket, { q: query, mode: newMode }))
     },
     [history, urls, bucket, query],
@@ -338,9 +333,6 @@ function Search({ bucket, query, page, mode, retry }) {
   )
 }
 
-// Possible values are undefined, 'objects', 'packages'
-const storage = mkStorage({ searchMode: 'SEARCH_MODE' })
-
 export default function BucketSearch({
   match: {
     params: { bucket },
@@ -352,22 +344,10 @@ export default function BucketSearch({
   const page = p && parseInt(p, 10)
   const retry = (params.retry && parseInt(params.retry, 10)) || undefined
 
-  const { urls } = NamedRoutes.use()
-  const history = useHistory()
-  React.useEffect(() => {
-    if (mode) return
-
-    const searchMode = storage.load()?.searchMode
-    switch (searchMode) {
-      case 'objects':
-      case 'packages':
-        history.replace(urls.bucketSearch(bucket, { q: query, mode: searchMode }))
-      // no default
-    }
-  }, [bucket, history, mode, query, urls])
-
   return (
     <M.Box pb={{ xs: 0, sm: 5 }} mx={{ xs: -2, sm: 0 }}>
+      <MetaTitle>{[query || 'Search', bucket]}</MetaTitle>
+
       {cfg ? (
         <Search {...{ bucket, query, page, mode, retry }} />
       ) : (