+++ catalog/app/website/pages/OpenLanding/OpenLanding.js
@@ -1,6 +1,7 @@
 import * as React from 'react'
 
 import * as LinkedData from 'utils/LinkedData'
+import MetaTitle from 'utils/MetaTitle'
 
 import Layout from 'website/components/Layout'
 import Contribute from 'website/components/Contribute'
@@ -14,6 +15,7 @@ import QuiltIsDifferent from './QuiltIsDifferent'
 export default function OpenLanding() {
   return (
     <Layout>
+      <MetaTitle />
       <React.Suspense fallback={null}>
         <LinkedData.CatalogData />
       </React.Suspense>